+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # OpenDR Toolkit Change Log
 
 ## Version 2.1.0
-Released on XX, XX, 2023.
+- New Features:
+    - Added YOLOv5 as an inference-only tool ([#360](https://github.com/opendr-eu/opendr/pull/360)).
 
 ## Version 2.0.0
 Released on December, 31st, 2022.